pyproject.toml
[project]
name = "agent-zero"
version = "0.1.0"
description = ""
authors = [
    {name = "Anupam Abhay",email = "anupamabhay1908@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.15" 

dependencies = [
    "langgraph (>=1.0.6,<2.0.0)",
    "langchain-google-genai (>=4.2.0,<5.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "duckduckgo-search (>=8.1.1,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "youtube-transcript-api (>=1.2.4,<2.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "gitpython (>=3.1.46,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
